+++ datadog/threadstats/__init__.py
@@ -1,1 +1,2 @@
 from datadog.threadstats.base import ThreadStats  # noqa
+from datadog.threadstats.aws_lambda import lambda_stats, datadog_lambda_wrapper  # noqa
